svr: txs: store tx in github too
