+++ server/data/github.test.js
@@ -2,7 +2,7 @@ const testData = require('../config/test').data
 const github = require('./github')
 const utils = require('./utils')
 
-// todo: git reset --hard head~xxx after tests / or just repo.deleteFile/updateFile
+// todo: git reset --hard head~xxx after tests / or just repo.deleteFile/updateFile in 'suiteSetup in test.js'
 // todo: do 50 upserts at the same time and see if auto merge conflict resolution works (test 2: create file and update at the same time)
 // todo: assert the changes with getContent
