+++ server/data/utils.js
@@ -5,3 +5,11 @@ exports.isCloseToDate = (date, closeToDate = new Date(), varianceSeconds = 30) =
   date.getTime() - varianceSeconds * 1000 < closeToDate.getTime()
 
 exports.getRandomInt = max => Math.floor(Math.random() * Math.floor(max))
+
+exports.getWeekNumber = date => {
+  const d = new Date(Date.UTC(date.getFullYear(), date.getMonth(), date.getDate()))
+  const dayNum = d.getUTCDay() || 7
+  d.setUTCDate(d.getUTCDate() + 4 - dayNum)
+  const yearStart = new Date(Date.UTC(d.getUTCFullYear(), 0, 1))
+  return Math.ceil((((d - yearStart) / 86400000) + 1) / 7)
+}
